Refactored starbeast parser and improved tests.
