+++ msprime/__init__.py
@@ -51,4 +51,4 @@ from msprime.simulations import *
 from msprime.exceptions import *
 from msprime.mutations import *
 from msprime.likelihood import *
-from msprime.species_tree import parse_species_tree, parse_starbeast
+from msprime.species_trees import parse_species_tree, parse_starbeast
