+++ tests/test_species_tree_parsing.py
@@ -22,6 +22,7 @@ Tests for the parsing of species trees in newick and StarBEAST format.
 import unittest
 
 import msprime
+import msprime.species_trees as species_trees
 
 
 def get_non_binary_tree(n):
@@ -36,6 +37,19 @@ def get_non_binary_tree(n):
             found = True
     assert found
     return tree
+
+
+class TestIsNumber(unittest.TestCase):
+    """
+    Test the is_number function.
+    """
+    def test_good_examples(self):
+        for x in ["2", "2.0", "1000", "-1e3", "1e-6"]:
+            self.assertTrue(species_trees.is_number(x))
+
+    def test_bad_examples(self):
+        for x in ["", "x2.0", "1000x", ";-1e3", ";;"]:
+            self.assertFalse(species_trees.is_number(x))
 
 
 class TestSpeciesTreeRoundTrip(unittest.TestCase):
@@ -56,7 +70,6 @@ class TestSpeciesTreeRoundTrip(unittest.TestCase):
             self.assertEqual(pop_config.initial_size, Ne)
             self.assertEqual(pop_config.growth_rate, 0)
             self.assertIn("species_name", pop_config.metadata)
-            self.assertTrue(pop_config.metadata["species_name"].isdigit())
 
         # Population IDs are mapped to leaves as they are encountered in a postorder
         # traversal.
@@ -68,6 +81,13 @@ class TestSpeciesTreeRoundTrip(unittest.TestCase):
                 k += 1
             else:
                 pop_id_map[u] = pop_id_map[tree.left_child(u)]
+
+        for u in tree.leaves():
+            pop_config = population_configurations[pop_id_map[u]]
+            self.assertEqual(pop_config.growth_rate, 0)
+            # Note: we're assuming the default newick here in tskit that labels
+            # nodes as their id + 1.
+            self.assertEqual(pop_config.metadata["species_name"], f"{u + 1}")
 
         # We should have demographic events for every non-unary internal node, and
         # events should be output in increasing time order.
@@ -227,10 +247,9 @@ class TestStarbeastRoundTrip(unittest.TestCase):
         if nexus is None:
             nexus = make_nexus(tree, pop_size_map)
         population_configurations, demographic_events = msprime.parse_starbeast(
-            nexus, branch_length_units, generation_time)
+            nexus, generation_time, branch_length_units)
         self.assertEqual(len(population_configurations), tree.num_samples())
         for pop_config in population_configurations:
-            # self.assertEqual(pop_config.initial_size, Ne)
             self.assertEqual(pop_config.growth_rate, 0)
             self.assertIn("species_name", pop_config.metadata)
 
@@ -247,9 +266,9 @@ class TestStarbeastRoundTrip(unittest.TestCase):
 
         for u in tree.leaves():
             pop_config = population_configurations[pop_id_map[u]]
-            # self.assertEqual(pop_config.initial_size, pop_size_map[u])
+            self.assertEqual(pop_config.initial_size, pop_size_map[u])
             self.assertEqual(pop_config.growth_rate, 0)
-            # TODO check the metadata: we should be encoding the species name
+            self.assertEqual(pop_config.metadata["species_name"], f"spc{u}")
 
         # We should have demographic events for every non-unary internal node, and
         # events should be output in increasing time order.
@@ -331,11 +350,12 @@ class TestSpeciesTreeParsingErrors(unittest.TestCase):
     """
     Tests for parsing of species trees in newick format.
     """
-    def test_bad_tree(self):
-        for bad_type in [None, {}, 123]:
-            with self.assertRaises(TypeError):
-                msprime.parse_species_tree(tree=bad_type, Ne=1)
+    def test_bad_params(self):
+        self.assertRaises(TypeError, msprime.parse_species_tree)
+        self.assertRaises(TypeError, msprime.parse_species_tree, tree="()")
+        self.assertRaises(TypeError, msprime.parse_species_tree, Ne=1)
 
+    def test_bad_tree(self):
         bad_trees = [
             "", ";",  "abcd", ";;;", "___", "∞",
             "(", ")", "()", "( )", "(()())",
@@ -361,7 +381,11 @@ class TestSpeciesTreeParsingErrors(unittest.TestCase):
                     Ne=good_ne,
                     generation_time=good_generation_time
                     )
-        for bad_ne in [None, -3]:
+
+        with self.assertRaises(TypeError):
+            msprime.parse_species_tree(good_tree, None)
+
+        for bad_ne in [-3, "x"]:
             with self.assertRaises(ValueError):
                 msprime.parse_species_tree(
                     good_tree,
@@ -369,7 +393,7 @@ class TestSpeciesTreeParsingErrors(unittest.TestCase):
                     Ne=bad_ne,
                     generation_time=good_generation_time
                     )
-        for bad_generation_time in [None, -3]:
+        for bad_generation_time in [None, -3, "x"]:
             with self.assertRaises(ValueError):
                 msprime.parse_species_tree(
                     good_tree,
@@ -386,7 +410,12 @@ class TestSpeciesTreeParsingErrors(unittest.TestCase):
                     generation_time=good_generation_time
                     )
 
-    def test_output(self):
+
+class TestSpeciesTreeExamples(unittest.TestCase):
+    """
+    Tests that we get the expected value in simple examples.
+    """
+    def test_4_species(self):
         good_tree = "(((human:5.6,chimpanzee:5.6):3.0,gorilla:8.6):9.4,orangutan:18.0)"
         good_branch_length_units = "myr"
         good_ne = 10000
@@ -398,14 +427,14 @@ class TestSpeciesTreeParsingErrors(unittest.TestCase):
                 generation_time=good_generation_time
                 )
         self.assertEqual(len(parsed_tuple), 2)
-        self.assertIs(type(parsed_tuple[0]), list)
+        self.assertIsInstance(parsed_tuple[0], list)
         self.assertEqual(len(parsed_tuple[0]), 4)
         for pc in parsed_tuple[0]:
-            self.assertIs(type(pc), msprime.simulations.PopulationConfiguration)
-        self.assertIs(type(parsed_tuple[1]), list)
+            self.assertIsInstance(pc, msprime.simulations.PopulationConfiguration)
+        self.assertIsInstance(parsed_tuple[1], list)
         self.assertEqual(len(parsed_tuple[1]), 3)
         for mm in parsed_tuple[1]:
-            self.assertIs(type(mm), msprime.simulations.MassMigration)
+            self.assertIsInstance(mm, msprime.simulations.MassMigration)
 
 
 class TestStarbeastParsingErrors(unittest.TestCase):
@@ -414,7 +443,7 @@ class TestStarbeastParsingErrors(unittest.TestCase):
     StarBEAST.
     """
     def test_bad_tree(self):
-        bad_trees = [None, {}, 123]
+        bad_trees = []
         tree_file = "tests/data/species_trees/101g_nucl_conc_unconst.combined.nwk.tre"
         with open(tree_file) as f:
             bad_trees.append(f.read())
@@ -467,6 +496,58 @@ class TestStarbeastParsingErrors(unittest.TestCase):
                         generation_time=good_generation_time
                         )
 
+    def test_bad_annotations(self):
+        good = "((1[&dmv={0.1}]:1,2[&dmv={0.2}]:1)[&dmv={0.3}])"
+        self.assertEqual(species_trees.strip_extra_annotations(good), good)
+        bad_examples = [
+            # No annotations
+            "((1:1,2:1)",
+            # Mismatched annotations
+            "((1[]:1,2[]:1)[]]",
+            "((1[]:1,2[]:1)[",
+            "((1[]:1,2[]:1)]",
+            # Missing all dmvs
+            "((1[]:1,2[]:1)[]",
+            # Missing closing }
+            "((1[&dmv={]:1,2[]:1)[]",
+        ]
+        for example in bad_examples:
+            with self.assertRaises(ValueError):
+                species_trees.strip_extra_annotations(example)
+
+    def test_bad_annotations_in_tree(self):
+        name_map = {f"{j}": f"{j}" for j in range(3)}
+        good = "(1[&dmv={1}]:1.14,2[&dmv={1}]:1.14)[&dmv={1}]"
+        pop_configs, demographic_events = species_trees.process_starbeast_tree(
+            good, 1, name_map)
+        self.assertEqual(len(pop_configs), 2)
+        self.assertEqual(len(demographic_events), 2)
+        bad_examples = [
+            # Missing one dmv
+            "(1[&dmv={1}]:1.14,2[&dmv={1}]:1.14)[&={1}]",
+            # No annotation
+            "(1[&dmv={1}]:1.14,2[&dmv={1}]:1.14)",
+        ]
+        for example in bad_examples:
+            with self.assertRaises(ValueError):
+                species_trees.process_starbeast_tree(example, 1, name_map)
+
+    def test_bad_translation(self):
+        good = "translate 1 spc1, 2 spc2, 3 spc3"
+        self.assertEqual(
+            species_trees.parse_translate_command(good),
+            {"1": "spc1", "2": "spc2", "3": "spc3"}
+        )
+        bad_examples = [
+           "translate 1,",
+           "translate 1 spc1 more, 2 spc2",
+           "translate 1 spc1, 1 spc2",
+           "translate 1 spc1, 2 spc1",
+        ]
+        for example in bad_examples:
+            with self.assertRaises(ValueError):
+                species_trees.parse_translate_command(example)
+
     def test_bad_parameter(self):
         with open("tests/data/species_trees/91genes_species_rev.tre") as f:
             good_tree = f.read()
@@ -477,7 +558,7 @@ class TestStarbeastParsingErrors(unittest.TestCase):
                             tree=f.read(),
                             branch_length_units=bad_branch_length_units,
                             generation_time=5)
-            for bad_generation_time in [-3]:
+            for bad_generation_time in [-3, "sdf"]:
                 with self.assertRaises(ValueError):
                     msprime.parse_starbeast(
                         tree=good_tree,
@@ -492,7 +573,12 @@ class TestStarbeastParsingErrors(unittest.TestCase):
                         generation_time=bad_generation_time
                         )
 
-    def test_output(self):
+
+class TestStarbeastExamples(unittest.TestCase):
+    """
+    Tests for known examples in starbeast format.
+    """
+    def test_12_species(self):
         with open("tests/data/species_trees/91genes_species_rev.tre") as f:
             good_tree = f.read()
             good_branch_length_units = "myr"
@@ -506,8 +592,11 @@ class TestStarbeastParsingErrors(unittest.TestCase):
             self.assertIs(type(parsed_tuple[0]), list)
             self.assertEqual(len(parsed_tuple[0]), 12)
             for pc in parsed_tuple[0]:
-                self.assertIs(type(pc), msprime.simulations.PopulationConfiguration)
-            self.assertIs(type(parsed_tuple[1]), list)
+                species_name = pc.metadata["species_name"]
+                self.assertTrue(species_name.startswith("spc"))
+                self.assertTrue(species_name[3:].isnumeric())
+                self.assertIsInstance(pc, msprime.simulations.PopulationConfiguration)
+            self.assertIsInstance(parsed_tuple[1], list)
             self.assertEqual(len(parsed_tuple[1]), 22)
             event_types = [msprime.simulations.MassMigration]
             event_types.append(msprime.simulations.PopulationParametersChange)
